feat: Add class AppError and  using asyncStorage

+++ src/storage/group/groupCreate.ts
@@ -1,16 +1,22 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { GROUP_COLLECTION } from "@storage/storageConfig";
+import { AppError } from "@utils/AppError";
+
 import { groupsGetAll } from "./groupsGetAll";
 
 export async function groupCreate(newGroup: string) {
-    try {
-        const storageGroups = await groupsGetAll();
+  try {
+    const storadGroups = await groupsGetAll();
 
-        const storage = JSON.stringify([...storageGroups, newGroup]);
-        await AsyncStorage.setItem(GROUP_COLLECTION, storage);
-        
-    } catch (error) {
-        throw error;
+    const groupAlreadyExistis = storadGroups.includes(newGroup);
+
+    if (groupAlreadyExistis) {
+      throw new AppError("Já existe um grupo cadastrado com esse nome.");
     }
 
+    const storage = JSON.stringify([...storadGroups, newGroup]);
+    await AsyncStorage.setItem(GROUP_COLLECTION, storage);
+  } catch (error) {
+    throw error;
+  }
 }
